change durationMs to unsigned long

+++ DurationFsm.h
@@ -8,11 +8,11 @@
 //define the functionality of the states
 class DurationState: public State {
   public:
-    DurationState(const unsigned int minDurationMs, String name): State(name), minDurationMs(minDurationMs), nextState(NULL) {
+    DurationState(const unsigned long minDurationMs, String name): State(name), minDurationMs(minDurationMs), nextState(NULL) {
     }
-    DurationState(const unsigned int minDurationMs, String name, State * const superState): State(name, superState), minDurationMs(minDurationMs), nextState(NULL) {
+    DurationState(const unsigned long minDurationMs, String name, State * const superState): State(name, superState), minDurationMs(minDurationMs), nextState(NULL) {
     }
-    const unsigned int minDurationMs;
+    const unsigned long minDurationMs;
     // nextState as NULL marks a state that is not changed when calling changeToNextStateIfElapsed(). minDurationMs is ignored in that case.
     DurationState *nextState;
 };
